views: some scaffolding
