+++ cpp/src/deck.gl/core/views/view-state.h
@@ -1,0 +1,3 @@
+class ViewState {
+  virtual ~ViewState() {};
+}
